// frontend/src/api/generated/models/fiscalSimulationStatusEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ACCEPTED` - ACCEPTED
 * * `WARNINGS` - WARNINGS
 * * `REJECTED` - REJECTED
 */
export type FiscalSimulationStatusEnum =
  (typeof FiscalSimulationStatusEnum)[keyof typeof FiscalSimulationStatusEnum];

export const FiscalSimulationStatusEnum = {
  ACCEPTED: "ACCEPTED",
  WARNINGS: "WARNINGS",
  REJECTED: "REJECTED",
} as const;